Delete Article now using UserService

+++ BlazingBlog.Application/Articles/DeleteArticle/DeleteArticleCommandHandler.cs
@@ -1,18 +1,25 @@
-﻿using BlazingBlog.Domain.Articles;
+﻿using BlazingBlog.Application.Users;
+using BlazingBlog.Domain.Articles;
 
 namespace BlazingBlog.Application.Articles.DeleteArticle
 {
     public class DeleteArticleCommandHandler : ICommandHandler<DeleteArticleCommand>
     {
         private readonly IArticleRepository _articleRepository;
+        private readonly IUserService _userService;
 
-        public DeleteArticleCommandHandler(IArticleRepository articleRepository)
+        public DeleteArticleCommandHandler(IArticleRepository articleRepository, IUserService userService)
         {
             _articleRepository = articleRepository;
+            _userService = userService;
         }
 
         public async Task<Result> Handle(DeleteArticleCommand request, CancellationToken cancellationToken)
         {
+            if (!await _userService.CurrentUserCanEditArticleAsync(request.Id))
+            {
+                return Result.Fail("You cannot delete this article! How did you get here?!");
+            }
             var deleted = await _articleRepository.DeleteArticleAsync(request.Id);
             if (deleted)
             {
